requirements.txt
aiobotocore==2.5.4
aiohttp==3.9.5
aioitertools==0.11.0
aiosignal==1.3.1
anndata==0.10.8
array_api_compat==1.7.1
asciitree==0.3.3
async-timeout==4.0.3
attrs==23.2.0
botocore==1.31.17
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmake==3.30.0
colorcet==3.1.0
contourpy==1.2.1
cycler==0.12.1
dask==2024.7.0
dask-expr==1.1.7
dask-image==2024.5.3
datashader==0.16.3
distributed==2024.7.0
docrep==0.3.2
easydl==2.2.0
einops==0.6.1
fasteners==0.19
filelock==3.15.4
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2023.6.0
geopandas==1.0.1
get-annotations==0.1.2
h5py==3.11.0
huggingface-hub==0.24.5
idna==3.7
igraph==0.11.6
imageio==2.34.2
importlib_metadata==8.0.0
importlib_resources==6.4.0
inflect==7.3.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.5
lazy_loader==0.4
legacy-api-wrap==1.4
leidenalg==0.10.2
lightning-utilities==0.11.5
lit==18.1.8
llvmlite==0.43.0
locket==1.0.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.2
matplotlib-scalebar==0.8.1
mdurl==0.1.2
more-itertools==10.3.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multipledispatch==1.0.0
multiscale_spatial_image==1.0.0
natsort==8.4.0
networkx==3.2.1
numba==0.60.0
numcodecs==0.12.1
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
ome-zarr==0.9.0
omnipath==1.0.8
opencv-python==4.8.0.76
pandas==2.0.3
param==2.1.1
partd==1.4.2
pathlib2==2.3.7.post1
patsy==0.5.6
Pillow==10.0.0
PIMS==0.7
pooch==1.8.2
pyarrow==17.0.0
pyct==0.5.0
pynndescent==0.5.13
pyogrio==0.9.0
pyparsing==3.0.9
pyproj==3.6.1
pytorch-lightning==2.0.8
pytz==2024.1
PyWavelets==1.6.0
PyYAML==6.0.1
requests==2.32.3
rich==13.7.1
s3fs==2023.6.0
safetensors==0.4.4
scanpy==1.10.2
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.13.1
scprep==1.2.3
seaborn==0.13.1
session_info==1.0.0
shapely==2.0.5
slicerator==1.1.0
sortedcontainers==2.4.0
spatial_image==1.1.0
spatialdata==0.2.1
statsmodels==0.14.2
stdlib-list==0.10.0
sympy==1.13.0
tblib==3.0.0
texttable==1.7.0
threadpoolctl==3.5.0
tifffile==2024.7.2
timm==1.0.8
tomli==2.0.1
toolz==0.12.1
torch==2.0.1
torchmetrics==1.4.0.post0
torchvision==0.15.2
tqdm==4.66.4
triton==2.0.0
typeguard==4.3.0
tzdata==2024.1
umap-learn==0.5.5
urllib3==1.26.19
validators==0.33.0
wrapt==1.16.0
xarray==2024.6.0
xarray-dataclasses==1.8.0
xarray-datatree==0.0.14
xarray-schema==0.0.3
xarray-spatial==0.4.0
yarl==1.9.4
zarr==2.18.2
zict==3.0.0
